Storing possible dead end
Committing to store code that was designed to solve the problem with the
regulator caring about a function of total services and private values.
Includes option to set number of periods payoffs continue for after
second and many other items. Likely a dead end because all land would be
converted in period 2. Not realistic.